o) refactoring: align AppState regex with the others. o) move SteamAppState to Data.Steam namespace
